Move pixel to its own file.
